added one route, moved private route
