+++ client/src/utils/API.js
@@ -1,0 +1,17 @@
+const TOKEN_KEY = 'jwt';
+
+export const login = () => {
+    localStorage.setItem(TOKEN_KEY, 'TestLogin');
+}
+
+export const logout = () => {
+    localStorage.removeItem(TOKEN_KEY);
+}
+
+export const isLogin = () => {
+    if (localStorage.getItem(TOKEN_KEY)) {
+        return true;
+    }
+
+    return false;
+}
