+++ package.json
@@ -1,4 +1,3 @@
-
 {
   "name": "mern",
   "version": "1.0.0",
@@ -27,3 +26,4 @@
     "if-env": "^1.0.4",
     "mongoose": "^5.12.13"
   }
+}
